RoboTraderBINANCE/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
dateparser==1.2.1
fastapi==0.115.14
Flask==3.1.1
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.5
pandas==2.2.3
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.7
pydantic_core==2.33.2
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.2
pytz==2025.2
regex==2024.11.6
requests==2.32.3
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinydb==4.8.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.0
